Format code with ruff and fix lint issues
- Update dependencies for Python 3.12+: numpy>=2.1, scipy>=1.14, pandas>=2.0
- Replace flake8/black/isort with ruff in dev dependencies
- Format all files with ruff formatter
- Fix lint issues: unused variables, line length, unnecessary int()
- Add noqa comments for pandas eval/query string variable usage
- Configure ruff to ignore E402 (import order) and RUF022 (__all__ order)

+++ openap/phase.py
@@ -1,8 +1,8 @@
 """Using fuzzy logic to identify flight phase in trajectory data."""
 
-import numpy as np
 from matplotlib import pyplot as plt
 
+import numpy as np
 from openap.extra import fuzzy
 
 
@@ -154,7 +154,7 @@ class FlightPhase:
             )
 
             state_raw = fuzzy.defuzz(self.states, aggregated, "lom")
-            state = int(round(state_raw))
+            state = round(state_raw)
             if state > 6:
                 state = 6
             if state < 1:
@@ -257,7 +257,6 @@ class FlightPhase:
         return (istart, ilof, iend + 1)
 
     def _get_fa_ld(self):
-
         # get the approach + landing data chunk (h=0)
         istart = 0
         iend = 0
@@ -318,7 +317,6 @@ class FlightPhase:
         return istart, iend
 
     def _get_de(self):
-
         labels = np.array(self.phaselabel())
 
         if "DE" not in labels:
